Fix error when user previous voice client is None

+++ neonbot/cogs/event.py
@@ -97,7 +97,7 @@ class Event(commands.Cog):
         )
 
         if player and player.connection:
-            if player.connection.is_paused() and members(after.channel.members) > 0:
+            if player.connection.is_paused() and after.channel and members(after.channel.members) > 0:
                 if player.messages.auto_paused:
                     await player.messages.auto_paused.delete()
                     player.messages.auto_paused = None
@@ -109,7 +109,7 @@ class Event(commands.Cog):
                 )
                 player.connection.resume()
             elif (
-                player.connection.is_playing() and members(before.channel.members) == 0
+                player.connection.is_playing() and before.channel and members(before.channel.members) == 0
             ):
                 msg = "Player paused because no users are listening."
                 log.cmd(member, msg, channel=before.channel, user="N/A")
